Changed MainClass Name / Moved ConnectionInfo
