+++ src/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: kr.ac.gachon.sw.closeheart.client.MainClass
+Main-Class: kr.ac.gachon.sw.closeheart.client.ClientMain
